requirements.txt
requests
beautifulsoup4
python-telegram-bot
python-dotenv
pycryptodome
